Fix sign up invitation link issue
